Final Version

+++ src/csd/aula2/rmi/SIFTBox.java
@@ -5,11 +5,9 @@ import java.io.BufferedReader;
 import java.io.File;
 import java.io.FileFilter;
 import java.io.FileInputStream;
-import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 import java.io.FileReader;
 import java.io.FileWriter;
-import java.io.FilenameFilter;
 import java.io.IOException;
 import java.rmi.Naming;
 import java.rmi.RemoteException;
@@ -18,25 +16,43 @@ import java.rmi.server.UnicastRemoteObject;
 
 import javax.rmi.ssl.SslRMIClientSocketFactory;
 import javax.rmi.ssl.SslRMIServerSocketFactory;
-import javax.swing.plaf.synth.SynthSeparatorUI;
 
-public class ServerImpl extends UnicastRemoteObject implements IServer {
+public class SIFTBox extends UnicastRemoteObject implements IServer {
+	/**
+	 * 
+	 */
+	private static final long serialVersionUID = 1L;
 	// public Map<Integer, File> map= new HashMap<Integer,File>();
 	private static String rootPath;
 	private static String myAddress;
 
-	protected ServerImpl() throws RemoteException {
+	protected SIFTBox() throws RemoteException {
 		super(9000, new SslRMIClientSocketFactory(), new SslRMIServerSocketFactory());
 
 	}
 
+	/**
+	 * Read the configuration file rootPath- path of the files myAddress -
+	 * server address (not used)
+	 */
 	private static void readConfigFile() throws IOException {
+		@SuppressWarnings("resource")
 		BufferedReader br = new BufferedReader(new FileReader("config.txt"));
 
 		rootPath = br.readLine();
 		myAddress = br.readLine();
 	}
 
+	/**
+	 * main - only responsible for starting the RMI server and invoke the
+	 * readConfigFile method
+	 * 
+	 * HOW TO RUN: java csd.aula2.rmi.SIFTBox On the config file there is the
+	 * address- do ifconfig the rootPath- our default test:
+	 * /home/osboxes/SIFTBoxServerDirs
+	 * 
+	 * @param args
+	 */
 	public static void main(String[] args) {
 		try {
 			System.setProperty("java.security.policy", "src/csd/aula2/rmi/policy.all");
@@ -50,20 +66,30 @@ public class ServerImpl extends UnicastRemoteObject implements IServer {
 				// do nothing - already started with rmiregistry
 			}
 
-			Naming.rebind("/myServer", new ServerImpl());
+			Naming.rebind("/siftBoxServer", new SIFTBox());
 			readConfigFile();
-			System.out.println("myServer bound in registry");
+			System.out.println("SIFTBox server secure RMI bound in registry");
 		} catch (Throwable th) {
 			th.printStackTrace();
 		}
 	}
 
+	/**
+	 * Method to put a new directory on Server Side Checks if the folder exists,
+	 * if it doesn't create a new folder with the .shares file Used for
+	 * starting, when the clients says which folders he wants to sync
+	 * 
+	 * @param directory
+	 *            - name of the directory to be created
+	 * @param clientId
+	 *            - client that made the request
+	 */
 	@Override
 	public void putDirectory(String directory, String clientId) throws RemoteException {
 		File f = new File(rootPath, directory);
 		if (!f.exists()) {
 			f.mkdir();
-
+			System.out.println("Creating directory "+ directory);
 			File shareFile = new File(f, ".shares");
 			try {
 				shareFile.createNewFile();
@@ -78,9 +104,21 @@ public class ServerImpl extends UnicastRemoteObject implements IServer {
 		}
 	}
 
+	/**
+	 * Method to put a file in a directory
+	 * 
+	 * @param directory
+	 *            - directory of the file
+	 * @param value
+	 *            - file that is going to be created on ServerSide
+	 * @param clientId
+	 *            - the id of the client that made the request to check the
+	 *            rights
+	 */
 	@Override
 	public void put(String directory, File value, String clientId) throws RemoteException, Exception {
 		if (hasRights(clientId, directory)) {
+			System.out.println("Creating file "+ value.getName() + " in directory "+directory);
 			byte[] data = new byte[Math.round(value.length())];
 			BufferedInputStream bis = new BufferedInputStream(new FileInputStream(value));
 			bis.read(data, 0, data.length);
@@ -99,25 +137,60 @@ public class ServerImpl extends UnicastRemoteObject implements IServer {
 		// map.put(key, value);
 	}
 
+	/**
+	 * File that return a file to the client
+	 * 
+	 * @param directory
+	 *            - name of the directory where the file is
+	 * @param name
+	 *            - name of the file
+	 * @param clientId
+	 *            - the id of the client that made the request to check the
+	 *            rights
+	 */
 	@Override
 	public File get(String directory, String name, String clientId) throws RemoteException {
 		if (hasRights(clientId, directory)) {
+			System.out.println("Getting file "+ name + " in directory "+directory);
 			File f = new File(rootPath + "/" + directory + name);
 			return f;
 		} else
 			return null;
 	}
 
+	/**
+	 * Method to delete a file on ServerSide
+	 * 
+	 * @param directory
+	 *            - directory where the file is
+	 * @param name
+	 *            - name of the file
+	 * @param clientId
+	 *            - the id of the client that made the request to check the
+	 *            rights
+	 */
 	@Override
 	public void delete(String directory, String name, String clientId) throws RemoteException {
 		File f = new File(rootPath + "/" + directory + name);
-		f.delete();
+		System.out.println("Deleting file "+ name + " in directory "+directory);
+		if (hasRights(clientId, directory) && f.exists()) {
+			f.delete();
+		}
 	}
 
+	/**
+	 * Method to get all the files in a given directory
+	 * 
+	 * @param directory
+	 *            - folder name where we will get all files
+	 * @param clientId
+	 *            - the id of the client that made the request to check the
+	 *            rights
+	 */
 	@Override
 	public File[] getAll(String directory, String clientId) throws RemoteException {
 		if (hasRights(clientId, directory)) {
-			System.out.println("I GOT ALL");
+			System.out.println("GETTING ALL FILES FROM: "+directory);
 			File f = new File(rootPath + "/" + directory);
 			if (f.isDirectory()) {
 				File[] temp = f.listFiles(new FileFilter() {
@@ -131,21 +204,26 @@ public class ServerImpl extends UnicastRemoteObject implements IServer {
 		return null;
 	}
 
-	@Override
-	public boolean checkDirectory(String directory) throws RemoteException {
-		File f = new File(rootPath + "/" + directory);
-		if (f.exists() && f.isDirectory()) {
-			return true;
-		} else
-			return false;
-	}
+	/*
+	 * @Override public boolean checkDirectory(String directory) throws
+	 * RemoteException { File f = new File(rootPath + "/" + directory); if
+	 * (f.exists() && f.isDirectory()) { return true; } else return false; }
+	 */
 
+	/**
+	 * Checks if a client has rights to a certain directory
+	 * 
+	 * @param clientId
+	 *            - the id of the client that made the request to check the
+	 *            rights
+	 * @param directory
+	 *            - directory to check if the client has right
+	 */
 	private boolean hasRights(String clientID, String directory) throws RemoteException {
 		boolean havePermissions = false;
 		File f = new File(rootPath + "/" + directory + "/.shares");
+		System.out.println("Checking permission on "+ directory+ " for client "+clientID );
 		if (f.exists() && f.isFile()) {
-			System.out.println(f.getName() + "   " + f.isHidden());
-			System.out.println(clientID + "   " + directory);
 			try {
 				FileReader fileReader = new FileReader(f);
 				BufferedReader bufferedReader = new BufferedReader(fileReader);
@@ -153,7 +231,6 @@ public class ServerImpl extends UnicastRemoteObject implements IServer {
 				while ((line = bufferedReader.readLine()) != null) {
 					if (line.equalsIgnoreCase(clientID)) {
 						havePermissions = true;
-						System.out.println(line + " " + line.equalsIgnoreCase(clientID));
 					}
 				}
 				bufferedReader.close();
